Chore(gmx): move path util to util.ts
