+++ packages/gmx/src/GMX.ts
@@ -5,24 +5,7 @@ import {
   CHAIN_ID_ARRAY
 } from './chain-ids.js'
 import { GMX_SWAP_ABI } from './abi.js'
-const buildPathQuery = (tokenIn?: string, tokenOut?: string) => {
-  // v3 paths are formatted as 0x<token><fee><token>
-
-  const conditions: FilterOperator[] = []
-
-  if (tokenIn) {
-    conditions.push({ $regex: `^${tokenIn}` })
-  }
-
-  if (tokenOut) {
-    // Chop the 0x prefix before comparing
-    conditions.push({ $regex: `${tokenOut.slice(2)}$` })
-  }
-
-  return {
-    $and: conditions,
-  }
-}
+import { buildPathQuery } from './utils.js'
 
 export const swap = async (swap: SwapActionParams) => {
   // This is the information we'll use to compose the Transaction object
