+++ packages/gmx/src/utils.ts
@@ -1,0 +1,18 @@
+export const buildPathQuery = (tokenIn?: string, tokenOut?: string) => {
+  // v3 paths are formatted as 0x<token><fee><token>
+
+  const conditions: FilterOperator[] = []
+
+  if (tokenIn) {
+    conditions.push({ $regex: `^${tokenIn}` })
+  }
+
+  if (tokenOut) {
+    // Chop the 0x prefix before comparing
+    conditions.push({ $regex: `${tokenOut.slice(2)}$` })
+  }
+
+  return {
+    $and: conditions,
+  }
+}
